requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

groq==0.4.2
openai==1.12.0
anthropic==0.18.0

chromadb==0.4.22
faiss-cpu==1.7.4
sentence-transformers==2.3.1
rank-bm25==0.2.2

pypdf==4.0.1
PyMuPDF==1.23.21
python-docx==1.1.0
openpyxl==3.1.2
pyyaml==6.0.1
python-magic==0.4.27
pillow==10.2.0
pytesseract==0.3.10

markdown==3.5.2
beautifulsoup4==4.12.3
lxml==5.1.0

numpy==1.26.3
scikit-learn==1.4.0
tiktoken==0.6.0

loguru==0.7.2
tenacity==8.2.3
tqdm==4.66.1
rich==13.7.0

streamlit==1.29.0